requirements.txt
amqp==2.5.2
asgiref==3.2.3
backcall==0.1.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.3.0
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0
